requirements.txt
dxcam
numpy<2
pillow